+++ src/lib/query.ts
@@ -1,6 +1,7 @@
-import { AxiosInstance } from '@contentstack/core';
+import { AxiosInstance, getData } from '@contentstack/core';
 import { BaseQuery } from './base-query';
-import { BaseQueryParameters, QueryOperation, QueryOperator, TaxonomyQueryOperation, params, queryParams } from './types';
+import { BaseQueryParameters, QueryOperation, QueryOperator, TaxonomyQueryOperation, params, queryParams, FindResponse } from './types';
+import { encodeQueryParams } from './utils';
 
 export class Query extends BaseQuery {
   private _contentTypeUid?: string;
@@ -593,5 +594,38 @@ export class Query extends BaseQuery {
     }
     this._parameters[key] = { '$gte': value };
     return this;
+  }
+
+  /**
+   * Override find method to include content type UID directly for better caching
+   */
+  override async find<T>(encode: boolean = false): Promise<FindResponse<T>> {
+    let requestParams: { [key: string]: any } = this._queryParams;
+
+    if (Object.keys(this._parameters).length > 0) {
+      let queryParams = { ...this._parameters };
+      
+      if (encode) {
+        queryParams = encodeQueryParams(queryParams);
+      }
+      
+      requestParams = { ...this._queryParams, query: queryParams };
+    }
+
+    const getRequestOptions: any = { 
+      params: requestParams,
+      // Add contentTypeUid directly for improved caching
+      contentTypeUid: this._contentTypeUid
+    };
+
+    if (this._variants) {
+      getRequestOptions.headers = {
+        ...getRequestOptions.headers,
+        'x-cs-variant-uid': this._variants
+      };
+    }
+    const response = await getData(this._client, this._urlPath, getRequestOptions);
+
+    return response as FindResponse<T>;
   }
 }